+++ test/preprocessor/DataLoader.py
@@ -2,7 +2,7 @@ import os; os.environ['XLA_FLAGS'] = os.environ.get('XLA_FLAGS', '') + ' --xla_f
 import jax; jax.config.update('jax_platforms', 'cpu')
 from pathlib import Path; import sys; sys.path.append(str(Path(__file__).resolve().parent.parent.parent))
 
-from lib.dataloader.DataLoader import DataLoader
+from lib.preprocessor.Preprocessor import Preprocessor
 from lib.dataset.dummy.load_dummy import load_dummy
 from lib.random.wrapper import seed2key
 
@@ -11,8 +11,8 @@ if __name__ == '__main__':
 
     sentences = load_dummy()
 
-    data_loader = DataLoader(sentences, key=key, batch_size_per_device=6, n_workers=8)
-    for n_batches, batch in enumerate(data_loader):
+    preprocessor = Preprocessor(sentences, key=key, batch_size_per_device=6, n_workers=8)
+    for n_batches, batch in enumerate(preprocessor):
         print(
             batch.src.shape,
             batch.dst.shape,
